[ARROW-5290][Java]Move the option to vector package
